Refactor index.js into each individual chart type
